+++ Practica1/Engine/src/main/java/com/example/engine/IGraphics.java
@@ -12,12 +12,12 @@ public interface IGraphics {
     public boolean changeBuffer();
     public void setResolution(float newX, float newY);
     public void translate (int x, int y);
-    public int realToLogicX(int x);
-    public int realToLogicY(int y);
-    public int realToLogicScale(int s);
-    public int logicToRealX(int x);
-    public int logicToRealY(int y);
-    public int logicToRealScale(int s);
+    public float realToLogicX(float x);
+    public float realToLogicY(float y);
+    public float realToLogicScale(float s);
+    public float logicToRealX(float x);
+    public float logicToRealY(float y);
+    public float logicToRealScale(float s);
     public void setColor(IColor color);
     public void setFont(IFont font);
 
@@ -32,15 +32,15 @@ public interface IGraphics {
     public void drawLine(float x, float y, float x_stop, float y_stop, IColor color);
     public void drawRectangle(float x, float y, float w, float h, IColor  color);
     public void fillRectangle(float x, float y, float w, float h,IColor color);
-    public void drawImage(IImage image, int x, int y, int w, int h);
+    public void drawImage(IImage image, float x, float y, float w, float h);
     public void drawText(String text, float x, float y, float textSize, IColor color);
     public void drawText(IFont font, String text, float x, float y, float textSize, IColor color);
     public void drawButton(IButton button);
 
     // Getters
-    public int getWidth();
-    public int getHeight();
-    public int getLogicWidth();
-    public int getLogicHeight();
+    public float getWidth();
+    public float getHeight();
+    public float getLogicWidth();
+    public float getLogicHeight();
     public float getScaleFactor();
 }
